pyproject.toml
[project]
name = "atlasmind"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.86.0",
    "fastmcp>=3.1.1",
    "httpx>=0.28.1",
    "python-dotenv>=1.2.2",
    "requests-oauthlib>=2.0.0",
]
